requirements.txt
Pillow==11.3.0
PyQt6==6.9.1
pyqt6_sip==13.10.2
Requests==2.32.5
screeninfo==0.8.1
